Progress bars

+++ src/main/kotlin/models/glTF/Buffer.kt
@@ -1,5 +1,5 @@
 package models.glTF
 
-class Buffer(filename: String, val byteLength: Int) {
+class Buffer(filename: String, val byteLength: Long) {
     val uri = "$filename.bin"
 }
